+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Output for plugin dir path
 - Avoid race conditions in signal forwarding
 - Ensure cleanup executed when para is interrupted
+- Avoid race conditions while reading files 
 
 ## 0.4.1 - 2019-07-19